Revert fixed frame commit, restructured folders
