+++ continuousControl/continuousControl.cpp
@@ -1,103 +1,312 @@
-#include <SDL2/SDL.h>
+#include <iostream>
+#include <fcntl.h>
+#include <unistd.h>
+#include <linux/input.h>
+#include <cmath>
 #include <thread>
 #include <atomic>
 #include <chrono>
 #include <csignal>
-#include <iostream>
+#include "../modules/custom/utilities/utils.h"
+#include <iomanip>
+#include "raisim/RaisimServer.hpp"
+#include "../modules/custom/arduinoConnection/arduinoController.h"
+#include "../modules/custom/rsTimedLoop/rsTimedLoop.h"
+#include "hexapod.h"
+#include <fstream>
+#include <memory>
 
-std::atomic<bool> running(true);
-std::atomic<float> inputAxisX(0.0f);
-std::atomic<float> inputAxisY(0.0f);
+
+using namespace std;
+
+const bool raisimSimulator = true;
+const float rsStep = 10; // Real Time Step (ms)
+
+RSTimedLoop rsLoop(rsStep);
+
+atomic<bool> running(true);
+atomic<float> inputAxisX(0.0f);
+atomic<float> inputAxisY(0.0f);
+atomic<bool> leftBumper(false);
+atomic<bool> rightBumper(false);
+atomic<bool> buttonA(false);
+atomic<bool> buttonB(false);
+atomic<bool> buttonY(false);
+atomic<bool> buttonX(false);
+atomic<int> dPadX(0);
+atomic<int> dPadY(0);
 
 void signalHandler(int signum) {
     running = false; // Stop the loops
 }
 
-void keyboardInputHandler() {
-    // Initialize SDL2 and create a window
-    if (SDL_Init(SDL_INIT_VIDEO) < 0) {
-        std::cerr << "Failed to initialize SDL: " << SDL_GetError() << std::endl;
+void readController(HexapodControl& hexapod) {
+    const char *device = "/dev/input/event20";  // Using event20 as specified
+    int fd = open(device, O_RDONLY);    
+    
+    if (fd == -1) {
+        cerr << "Failed to open input device." << endl;
         return;
     }
 
-    SDL_Window* window = SDL_CreateWindow("SDL2 Keyboard Input",
-                                          SDL_WINDOWPOS_UNDEFINED,
-                                          SDL_WINDOWPOS_UNDEFINED,
-                                          640, 480,
-                                          SDL_WINDOW_SHOWN);
-    if (!window) {
-        std::cerr << "Failed to create window: " << SDL_GetError() << std::endl;
-        SDL_Quit();
-        return;
-    }
-
-    SDL_Event event;
+    struct input_event ev;
 
     while (running) {
-        while (SDL_PollEvent(&event)) {
-            if (event.type == SDL_QUIT) {
-                running = false;
+        ssize_t n = read(fd, &ev, sizeof(ev));
+        if (n == (ssize_t)-1) {
+            cerr << "Failed to read input event." << endl;
+            break;
+        }
+
+        if (ev.type == EV_ABS || ev.type == EV_KEY) {
+            // Handle deadzone for Right Trigger (code 9) and Left Trigger (code 10)
+            if ((ev.code == 9 || ev.code == 10) && abs(ev.value) < 2048) {
+                continue; // Skip further processing if within the deadzone
             }
-            if (event.type == SDL_KEYDOWN || event.type == SDL_KEYUP) {
-                float value = (event.type == SDL_KEYDOWN) ? 1.0f : 0.0f;
 
-                switch (event.key.keysym.sym) {
-                    case SDLK_w: // Forward
-                        inputAxisY = inputAxisY-value;
-                        break;
-                    case SDLK_s: // Backward
-                        inputAxisY = inputAxisY+value;
-                        break;
-                    case SDLK_a: // Left
-                        inputAxisX = inputAxisX-value;
-                        break;
-                    case SDLK_d: // Right
-                        inputAxisX = inputAxisX+value;
-                        break;
-                    case SDLK_ESCAPE: // Exit on Escape
-                        running = false;
-                        break;
-                    default:
-                        break;
-                }
+            switch (ev.code) {
+                case 0: // Left Joystick X
+                    // cout << "Left Joystick X: " << ev.value << " (" 
+                    //           << (ev.value < 0 ? "Left" : "Right") << ")" << endl;
 
-                // Print which key was pressed or released
-                // std::cout << "Key event: " << SDL_GetKeyName(event.key.keysym.sym)
-                //           << " | X: " << inputAxisX.load()
-                //           << " | Y: " << inputAxisY.load() << std::endl;
+                    inputAxisX = ev.value;
+                    break;
+                case 1: // Left Joystick Y
+                    // cout << "Left Joystick Y: " << ev.value << " (" 
+                    //           << (ev.value < 0 ? "Up" : "Down") << ")" << endl;
+
+                    inputAxisY = ev.value;
+                    break;
+                // case 2: // Right Joystick X
+                //     // cout << "Right Joystick X: " << ev.value << " (" 
+                //     //           << (ev.value < 0 ? "Left" : "Right") << ")" << endl;
+                //     break;
+                // case 5: // Right Joystick Y
+                //     // cout << "Right Joystick Y: " << ev.value << " (" 
+                //     //           << (ev.value < 0 ? "Up" : "Down") << ")" << endl;
+                //     break;
+                case 16: // Dpad X
+                    // cout << "Dpad X: " << ev.value << " (" 
+                    //           << (ev.value < 0 ? "Left" : "Right") << ")" << endl;
+                    dPadX = ev.value;
+                    break;
+                case 17: // DPad Y
+                    // cout << "DPad Y: " << ev.value << " (" 
+                    //           << (ev.value > 0 ? "Down" : "Up") << ")" << endl;
+                    dPadY = ev.value;
+                    break;
+                // case 9: // Right Trigger
+                //     // cout << "Right Trigger: " << ev.value << endl;
+                //     break;
+                // case 10: // Left Trigger
+                //     // cout << "Left Trigger: " << ev.value << endl;
+                //     break;
+                // case 317: // Left Joystick Click
+                //     // cout << "Left Joystick Click: " << ev.value << " (" 
+                //     //           << (ev.value == 1 ? "Pressed" : "Released") << ")" << endl;
+                //     break;
+                // case 318: // Right Joystick Click
+                //     // cout << "Right Joystick Click: " << ev.value << " (" 
+                //     //           << (ev.value == 1 ? "Pressed" : "Released") << ")" << endl;
+                //     break;
+                case 304: // Button A
+                    // cout << "Button A: " << ev.value << " (" 
+                    //           << (ev.value == 1 ? "Pressed" : "Released") << ")" << endl;
+                    buttonA = ev.value;
+                    break;
+                case 307: // Button X
+                    // cout << "Button X: " << ev.value << " (" 
+                    //           << (ev.value == 1 ? "Pressed" : "Released") << ")" << endl;
+                    buttonX = ev.value;
+                    break;
+                case 305: // Button B
+                    static bool wasButtonBPressed = false; // Track the previous state of Button Y
+                    // cout<<"ButtonB Pressed: " <<endl;
+                    if (ev.value == 1 && !wasButtonBPressed) {
+                        // Button B is pressed and was not pressed before (i.e., transition from unpressed to pressed)
+                        cout<<"ButtonB Pressed: " <<endl;
+                        
+                        buttonB = true;
+                        
+                        hexapod.operationDuration = 0;
+                        
+                        wasButtonBPressed = true; // Update the state to pressed
+                    } else {
+                        // Button B is released, reset the state
+                        buttonB = false;
+                        wasButtonBPressed = false;
+                    }
+                    break;
+                case 308: // Button Y
+                    static bool wasButtonYPressed = false; // Track the previous state of Button Y
+                    
+                    if (ev.value == 1 && !wasButtonYPressed) {
+                        // Button Y is pressed and was not pressed before (i.e., transition from unpressed to pressed)
+                        buttonY = true;
+                        
+                        // Output the desired information
+                        cout << "Current Angles: " << hexapod.currentAngles << endl;
+                        cout << "Desired Angles: " << hexapod.desiredAngles << endl;
+                        hexapod.printPos();
+                        cout <<endl;
+                        
+                        wasButtonYPressed = true; // Update the state to pressed
+                    } else {
+                        // Button Y is released, reset the state
+                        buttonY = false;
+                        wasButtonYPressed = false;
+                    }
+                    break;
+                case 311: // Right Bumper
+                    // cout << "Right Bumper: " << ev.value << " (" 
+                    //           << (ev.value == 1 ? "Pressed" : "Released") << ")" << endl;
+                    
+                    
+                    rightBumper = ev.value;
+                    break;
+                case 310: // Left Bumper
+                    // cout << "Left Bumper: " << ev.value << " (" 
+                    //           << (ev.value == 1 ? "Pressed" : "Released") << ")" << endl;
+                    leftBumper = ev.value;
+                    break;
+                default:
+                    // cout << "Unknown event. Code: " << ev.code << " Value: " << ev.value << endl;
+                    break;
+            }
+        }
+    }
+
+    close(fd);
+}
+
+void runHexapod(HexapodControl& hexapod) {
+    
+    // Record start time
+    auto startTime = chrono::high_resolution_clock::now();
+    hexapod.desiredAngles = hexapod.simulator->convertVecDynToEigen(hexapod.simulator->hexapodLegModel->getGeneralizedCoordinate());
+    // cout<<hexapod.desiredAngles;
+    hexapod.rsLoop.updateTimeDelay();
+
+    while (running) {
+        // hexapod.rsLoop.realTimeDelay();
+        // hexapod.simulator->server->integrateWorldThreadSafe();
+        // continue;
+        // Get current position
+        hexapod.currentAngles = hexapod.simulator->convertVecDynToEigen(hexapod.simulator->hexapodLegModel->getGeneralizedCoordinate());
+        hexapod.currentAngularVelocities = hexapod.simulator->convertVecDynToEigen(hexapod.simulator->hexapodLegModel->getGeneralizedVelocity());
+
+        hexapod.updatePos();
+        
+        // Preset holding position
+        hexapod.simulator->setSimVelocity(hexapod.desiredAngles, Eigen::VectorXd::Zero(18));
+
+        if (hexapod.active) {
+            float moveVectorMag = Utils::constrain(sqrt(pow(inputAxisX, 2) + pow(inputAxisY, 2)), 0, 2047)   / 2047 * 100;
+            float moveVectorAng = atan2(-inputAxisY, inputAxisX);
+
+            // Remap the angle from [-π, π] to [0, 2π]
+            if (moveVectorAng < 0) {
+                moveVectorAng += 2 * M_PI;
+            }
+
+            // Set precision and width for consistent output
+            cout << "\rMove Vector: " 
+                    << fixed << setprecision(2) << setw(6) << moveVectorMag << "\% at "
+                    << fixed << setprecision(2) << setw(6) << moveVectorAng 
+                    << flush;
+        }
+        
+        // hexapod.moveToOff();
+
+        if (leftBumper && rightBumper) {
+
+            if (!hexapod.active) {
+                cout<<endl<<"standing" <<endl;
+                hexapod.active = true;
+                hexapod.stand();
+                
+            } else {
+                cout<<endl<<"turning off" <<endl;
+                hexapod.active = false;
+                hexapod.off();
+                
+            }
+           
+        } 
+        
+
+        // if (buttonA) {
+        //     hexapod.moveToZero();
+        // }
+
+        if (buttonX) {
+            if (dPadX == -1) {
+                hexapod.jacobianTest(0);
+            } else if (dPadX == 1) {
+                hexapod.jacobianTest(1);
+            } else if (dPadY == 1) {
+                hexapod.jacobianTest(2);
             }
         }
 
-        std::this_thread::sleep_for(std::chrono::milliseconds(5));
+        // if (buttonY) {
+        //     cout<<"Current Angles: "<< hexapod.currentAngles<<endl;
+        //     cout<<"Desired Angles: "<< hexapod.desiredAngles<<endl;
+        //     hexapod.printPos();
+        // }
+
+        hexapod.rsLoop.realTimeDelay();
+        hexapod.simulator->server->integrateWorldThreadSafe();
     }
 
-    SDL_DestroyWindow(window);
-    SDL_Quit();
+    // Calculate the duration
+    auto endTime = chrono::high_resolution_clock::now();
+    chrono::duration<double> elapsed = endTime - startTime;
+
+    hexapod.simulator->server->killServer();
+    
+    // Output the duration
+    cout << "\nrunHexapod loop duration: " << elapsed.count() << " seconds." << endl;
 }
 
-int main() {
+int main(int argc, char* argv[]) {
+
+    // Set the path for raisim simulator
+    Path binaryPath = raisim::Path::setFromArgv(argv[0]);
+
+    // Setup communication file stream for arduino connection
+    fstream arduinoPort("/dev/ttyACM0");
+
+    // Create arduino controller as blank unique pointer
+    unique_ptr<ArduinoController> arduino;
+
+    // Assume arduino Connected
+    bool arduinoConnected = true;
+
+    if (arduinoPort) {
+        // Set the Arduino Controller up with port and abud rate
+        arduino = make_unique<ArduinoController>("/dev/ttyACM0", 921600);
+
+        cout << "Arduino connected." << endl;
+        
+    } else {
+        // Set arduino as a blank controller
+        arduino = make_unique<ArduinoController>();
+
+        // Change sim mode
+        arduinoConnected = false;
+
+        cout << "Arduino not connected. Running in simulation mode." << endl;
+    }
+    // Create Hexapod 
+    HexapodControl hexapod(1, move(arduino), rsLoop, arduinoConnected, raisimSimulator, rsStep, binaryPath);
+    // Start the controller input reading in a separate thread
+    thread input_thread(readController, ref(hexapod));
     // Register signal handler
     signal(SIGINT, signalHandler);
 
-    // Start the keyboard input handler thread
-    std::thread keyboardThread(keyboardInputHandler);
-
-    // Main loop: Print X and Y axis values
-    while (running) {
-        float axisX = inputAxisX.load();
-        float axisY = inputAxisY.load();
-
-        std::cout
-                          << " | X: " << axisX
-                          << " | Y: " << axisY << std::endl;
-
-
-        // Sleep for 100 ms to reduce console spam
-        std::this_thread::sleep_for(std::chrono::milliseconds(5));
-    }
-
-    // Stop the keyboard input handler thread
-    keyboardThread.join();
+    // Run the simulation in the main thread
+    runHexapod(hexapod);
 
     return 0;
 }
