+++ continuousControl/hexapod.cpp
@@ -1,7 +1,7 @@
 #include "hexapod.h"
 #include "../modules/custom/rsTimedLoop/rsTimedLoop.h"
 #include "../modules/custom/utilities/utils.h"
-#include "../modules/custom/raisimSimulatorLeg/raisimSimulator.h"
+#include "../modules/custom/raisimSimulatorFull/raisimSimulator.h"
 
 #include "matplotlibcpp.h"
 
@@ -23,13 +23,17 @@ HexapodControl::HexapodControl(unsigned int id, std::unique_ptr<ArduinoControlle
 {
     if (raisimSimulator)
     {
-        simulator = make_unique<RaisimSimulator>(rsStep, binaryPath, "HexapodControlV2b.urdf");
+        simulator = make_unique<RaisimSimulator>(rsStep, binaryPath, "hexapodCustom.urdf");
     }
 
     currentAngles = Eigen::VectorXd(18);
     pos = Eigen::VectorXd(18);
+    desiredAngles = Eigen::VectorXd(18);
+    currentAngularVelocities = Eigen::VectorXd(18);
 
-    // moveToCurled();
+    active = false;
+
+    moveToCurled();
 }
 
 HexapodControl::~HexapodControl()
@@ -56,21 +60,35 @@ Eigen::MatrixXd HexapodControl::getJacobian(int legNum) const
     Jac(2, 1) = tibiaX*cos(currentAngles[legNum*3+1] + currentAngles[legNum*3+2]) + femurX*cos(currentAngles[legNum*3+1]);
     Jac(2, 2) = tibiaX*cos(currentAngles[legNum*3+1] + currentAngles[legNum*3+2]);
 
-    // cout << "Jac" << endl << Jac << endl;
-
     return Jac;
 }
 
 // Perform forward kinematics of the HexapodControl leg to get end affector positions
 void HexapodControl::updatePos()
 {
+    // for (int legNum = 0; legNum < 6; legNum++) {
+    
+    //     pos(legNum*3) = coxaX*cos(bodyLegAngles[legNum] + currentAngles[legNum*3+0]) + bodyLegOffsets[legNum]*cos(bodyLegAngles[legNum]) + tibiaX*cos(bodyLegAngles[legNum] + currentAngles[legNum*3+0])*cos(currentAngles[legNum*3+1] + currentAngles[legNum*3+2]) + femurX*cos(bodyLegAngles[legNum] + currentAngles[legNum*3+0])*cos(currentAngles[legNum*3+1]);
+    //     pos(legNum*3+1) = coxaX*sin(bodyLegAngles[legNum] + currentAngles[legNum*3+0]) + bodyLegOffsets[legNum]*sin(bodyLegAngles[legNum]) + tibiaX*cos(currentAngles[legNum*3+1] + currentAngles[legNum*3+2])*sin(bodyLegAngles[legNum] + currentAngles[legNum*3+0]) + femurX*sin(bodyLegAngles[legNum] + currentAngles[legNum*3+0])*cos(currentAngles[legNum*3+1]);
+    //     pos(legNum*3+2) = coxaZ + tibiaX*sin(currentAngles[legNum*3+1] + currentAngles[legNum*3+2]) + femurX*sin(currentAngles[legNum*3+1]);
+    // }
+
+    pos = doFK(currentAngles);
+}
+
+Eigen::VectorXd HexapodControl::doFK(Eigen::VectorXd angs) {
+    Eigen::VectorXd position(18);
+
     for (int legNum = 0; legNum < 6; legNum++) {
     
-        pos(legNum*3) = coxaX*cos(bodyLegAngles[legNum] + currentAngles[legNum*3+0]) + bodyLegOffsets[legNum]*cos(bodyLegAngles[legNum]) + tibiaX*cos(bodyLegAngles[legNum] + currentAngles[legNum*3+0])*cos(currentAngles[legNum*3+1] + currentAngles[legNum*3+2]) + femurX*cos(bodyLegAngles[legNum] + currentAngles[legNum*3+0])*cos(currentAngles[legNum*3+1]);
-        pos(legNum*3+1) = coxaX*sin(bodyLegAngles[legNum] + currentAngles[legNum*3+0]) + bodyLegOffsets[legNum]*sin(bodyLegAngles[legNum]) + tibiaX*cos(currentAngles[legNum*3+1] + currentAngles[legNum*3+2])*sin(bodyLegAngles[legNum] + currentAngles[legNum*3+0]) + femurX*sin(bodyLegAngles[legNum] + currentAngles[legNum*3+0])*cos(currentAngles[legNum*3+1]);
-        pos(legNum*3+2) = coxaZ + tibiaX*sin(currentAngles[legNum*3+1] + currentAngles[legNum*3+2]) + femurX*sin(currentAngles[legNum*3+1]);
+        position(legNum*3) = coxaX*cos(bodyLegAngles[legNum] + angs[legNum*3+0]) + bodyLegOffsets[legNum]*cos(bodyLegAngles[legNum]) + tibiaX*cos(bodyLegAngles[legNum] + angs[legNum*3+0])*cos(angs[legNum*3+1] + angs[legNum*3+2]) + femurX*cos(bodyLegAngles[legNum] + angs[legNum*3+0])*cos(angs[legNum*3+1]);
+        position(legNum*3+1) = coxaX*sin(bodyLegAngles[legNum] + angs[legNum*3+0]) + bodyLegOffsets[legNum]*sin(bodyLegAngles[legNum]) + tibiaX*cos(angs[legNum*3+1] + angs[legNum*3+2])*sin(bodyLegAngles[legNum] + angs[legNum*3+0]) + femurX*sin(bodyLegAngles[legNum] + angs[legNum*3+0])*cos(angs[legNum*3+1]);
+        position(legNum*3+2) = coxaZ + tibiaX*sin(angs[legNum*3+1] + angs[legNum*3+2]) + femurX*sin(angs[legNum*3+1]);
     }
+
+    return position;
 }
+
 // Print End Affector Positions
 void HexapodControl::printPos() const
 {
@@ -108,8 +126,11 @@ void HexapodControl::moveToOff()
     for (int i = 0; i < 18; ++i) {
         offAnglesVector[i] = offAngles[i % 3]; // Repeat the set of 3 angles
     }
+
     setAngs(offAnglesVector);
+    // moveLegsToPos(offAnglesVector);
 }
+
 // Move to curled position
 void HexapodControl::moveToCurled()
 {
@@ -118,17 +139,48 @@ void HexapodControl::moveToCurled()
     for (int i = 0; i < 18; ++i) {
         offAnglesVector[i] = offAngles[i % 3]; // Repeat the set of 3 angles
     }
+
     setAngs(offAnglesVector);
+}
+
+// Move to initial position for walk cycle
+void HexapodControl::off()
+{
+
+    Eigen::VectorXd legOffPos(18);
+
+    double tempPosA[18] = {
+        0.235586, 0.362771, 0.0,
+        0.382555, 0.0, 0.0,
+        0.235586, -0.362771, 0.0,
+        -0.235586, -0.362771, 0.0,
+        -0.382555, 0.0, 0.0,
+        -0.235586, 0.362771, 0.0
+    };
+
+    legOffPos = Eigen::Map<Eigen::VectorXd>(tempPosA, 18);
+    moveLegsToPos(legOffPos);
+
+    double tempPosB[18] = {
+        0.19, 0.29, 0.0,
+        0.29, 0, 0.0,
+        0.19, -0.29, 0.0,
+        -0.19, -0.29, 0.0,
+        -0.29, 0, 0.0,
+        -0.19, 0.29, 0.0
+    };
+
+    legOffPos = Eigen::Map<Eigen::VectorXd>(tempPosB, 18);
+    moveLegsToPos(legOffPos);
 }
 
 // Move to initial position for walk cycle
 void HexapodControl::stand()
 {
-    Eigen::VectorXd nextAngles(18);
 
     Eigen::VectorXd legstandPos(18);
 
-    double tempPos[18] = {
+    double tempPosA[18] = {
         0.235586, 0.362771, 0,
         0.382555, 0, 0,
         0.235586, -0.362771, 0,
@@ -137,12 +189,10 @@ void HexapodControl::stand()
         -0.235586, 0.362771, 0
     };
 
-    legstandPos = Eigen::Map<Eigen::VectorXd>(tempPos, 18);
-
+    legstandPos = Eigen::Map<Eigen::VectorXd>(tempPosA, 18);
     moveLegsToPos(legstandPos);
 
-    // Update tempPos with new values
-    double tempPos2[18] = {
+    double tempPosB[18] = {
         0.235586, 0.362771, -0.113248,
         0.382555, 0, -0.113248,
         0.235586, -0.362771, -0.113248,
@@ -151,12 +201,12 @@ void HexapodControl::stand()
         -0.235586, 0.362771, -0.113248
     };
 
-    legstandPos = Eigen::Map<Eigen::VectorXd>(tempPos2, 18);
-
+    legstandPos = Eigen::Map<Eigen::VectorXd>(tempPosB, 18);
     moveLegsToPos(legstandPos);
 }
 
 void HexapodControl::moveLegsToPos(const Eigen::VectorXd& desiredPos) {
+
 
     Eigen::Vector3d desiredSpatialVelocity;
     Eigen::Vector3d legJointVelocity;
@@ -165,46 +215,43 @@ void HexapodControl::moveLegsToPos(const Eigen::VectorXd& desiredPos) {
     Eigen::MatrixXd jacobian;
     Eigen::MatrixXd jacobianPseudoInverse;
 
-
     Eigen::VectorXd posOffset(18);
     posOffset = desiredPos - pos;
-
     int dur = 1000;
 
-    // cout<<endl<< "Leg " << legNum <<endl;
-    // cout<< "currentPos: " << currentPos.transpose() <<endl;
+    // cout<< "currentPos: " << pos.transpose() <<endl;
     // cout<< "desiredPos: " << desiredPos.transpose() <<endl;
     // cout<< "posOffset: " << posOffset.transpose() <<endl;
-    // cout<< "distance: " << dis <<endl;
     // cout<< "desiredSpatialVelocity: " << desiredSpatialVelocity.transpose() <<endl;
     // cout<< "legJointVelocity: " << legJointVelocity.transpose() <<endl;
 
 
     // Update time for real time loop
-    rsLoop.updateTimeDelay();
-
-    for (int i = 0; i <= dur / rsStep; i++) {
-
+    // rsLoop.updateTimeDelay();
+    for (int i = 0; i < dur / rsStep; i++) {
         desiredAngularVelocities*=0;
 
         for (int legNum = 0; legNum < 6; legNum++) {
-
             desiredSpatialVelocity << posOffset(legNum*3), posOffset(legNum*3+1), posOffset(legNum*3+2);
-
             jacobian = getJacobian(legNum);
             jacobianPseudoInverse = jacobian.completeOrthogonalDecomposition().pseudoInverse();
             legJointVelocity = jacobianPseudoInverse * (desiredSpatialVelocity / (dur/1000));
-
             for (int joint = 0; joint < 3; joint++) {
                 desiredAngularVelocities[legNum*3+joint] = legJointVelocity[joint]; // Repeat the set of 3 angles
             }
         }
 
         nextAngles = currentAngles + desiredAngularVelocities * (rsStep / 1000);
+        simulator->setSimVelocity(nextAngles, desiredAngularVelocities);
 
-        setAngs(nextAngles);
+        currentAngles = nextAngles;
+        updatePos();
+
+        simulator->server->integrateWorldThreadSafe();
         rsLoop.realTimeDelay();
     }
+
+    desiredAngles = currentAngles;
     
 }
 
@@ -274,6 +321,7 @@ void HexapodControl::moveLegToPos(const Eigen::Vector3d& desiredPos, const int l
 void HexapodControl::setAngs(const Eigen::VectorXd &angs)
 {
     currentAngles = angs;
+    
     sendAngs();
 
     updatePos();
@@ -363,13 +411,120 @@ void HexapodControl::sendAngs()
     if (simulator) {
 
         // Subtract the values at the specified indices from 180
-        modifiedAngs(10) = - modifiedAngs(10);
-        modifiedAngs(11) = - modifiedAngs(11);
-        modifiedAngs(13) = - modifiedAngs(13);
-        modifiedAngs(14) = - modifiedAngs(14);
-        modifiedAngs(16) = - modifiedAngs(16);
-        modifiedAngs(17) = - modifiedAngs(17);
+        // modifiedAngs(10) = - modifiedAngs(10);
+        // modifiedAngs(11) = - modifiedAngs(11);
+        // modifiedAngs(13) = - modifiedAngs(13);
+        // modifiedAngs(14) = - modifiedAngs(14);
+        // modifiedAngs(16) = - modifiedAngs(16);
+        // modifiedAngs(17) = - modifiedAngs(17);
 
         simulator->setSimAngle(modifiedAngs);
+
+        desiredAngles = modifiedAngs;
     }
 }
+
+void HexapodControl::jacobianTest(const int &style)
+{
+    // Test Control Variables
+    float radius = 0.06; // meters
+    double period = 2;   // secs
+    double cycles = 5;
+
+    // Find Duration
+    int dur = period * cycles * 1000; // ms
+
+    // Create vector variables for control calculations
+    Eigen::Vector3d desiredSpatialVelocity;
+    Eigen::Vector3d legJointVelocity;
+    Eigen::VectorXd desiredAngularVelocities(18);
+    Eigen::VectorXd nextAngles(18);
+    // Eigen::Vector3d nextPos;
+    Eigen::MatrixXd jacobian;
+    Eigen::MatrixXd jacobianPseudoInverse;
+
+    // // Set Start Position
+    // float offAngles[3] = {0, 40 * M_PI / 180, 258 * M_PI / 180};
+    // for (int i = 0; i < 18; ++i) {
+    //     nextAngles[i] = offAngles[i % 3]; // Repeat the set of 3 angles
+    // }
+    // setAngs(nextAngles);
+
+    operationDuration = dur / rsStep + 2000/rsStep;
+
+    // Simulation Cycle
+    for (int i = 0; i <= dur / rsStep + 2000/rsStep; i++)
+    {
+        cout<<operationDuration<<endl;
+        if (operationDuration <=0) {
+            cout<<"exit"<<endl;
+            break;
+        }
+
+        if (i < 2000/rsStep) {
+            simulator->setSimVelocity(desiredAngles, Eigen::VectorXd::Zero(18));
+            rsLoop.realTimeDelay();
+            // Integrate the Simulator server 
+            simulator->server->integrateWorldThreadSafe();
+            operationDuration--;
+            continue;
+        }
+
+        // desiredSpatialVelocity << 0,0,0;
+        // Desired spatial velocity of XYZ
+
+        for (int legNum = 0; legNum < 6; legNum++) {
+
+            if (style == 0) {
+                // cout<<endl<<"Jacobian test in X-Y plane" <<endl;
+                desiredSpatialVelocity << 
+                    0,
+                    radius * 2 / period * M_PI * cos(2 / period * M_PI * (i) * (rsStep / 1000)),
+                    -radius * 2 / period * M_PI * sin(2 / period * M_PI * (i) * (rsStep / 1000));
+            } else if (style == 1) {
+                // cout<<endl<<"Jacobian test in X-Z plane" <<endl;
+                desiredSpatialVelocity << 
+                    radius * 2 / period * M_PI * cos(2 / period * M_PI * (i) * (rsStep / 1000)),
+                    0,
+                    -radius * 2 / period * M_PI * sin(2 / period * M_PI * (i) * (rsStep / 1000));
+            } else if (style == 2) {
+                // cout<<endl<<"Jacobian test in Y-Z plane" <<endl;
+                desiredSpatialVelocity << 
+                    radius * 2 / period * M_PI * cos(2 / period * M_PI * (i) * (rsStep / 1000)),
+                    radius * 2 / period * M_PI * sin(2 / period * M_PI * (i) * (rsStep / 1000)),
+                    0;
+            }
+        
+            jacobian = getJacobian(legNum);
+            jacobianPseudoInverse = jacobian.completeOrthogonalDecomposition().pseudoInverse();
+            legJointVelocity = jacobianPseudoInverse * desiredSpatialVelocity;
+
+            for (int joint = 0; joint < 3; joint++) {
+                desiredAngularVelocities[legNum*3+joint] = legJointVelocity[joint]; // Repeat the set of 3 angles
+            }
+
+        }
+
+        // Find next aqngles using discrete integration
+        nextAngles = currentAngles + desiredAngularVelocities * (rsStep / 1000);
+
+        // Send the angles to the arduino and simulation as desired        
+        // setAngs(nextAngles);
+
+        // Send Velcoties to the simulator
+        simulator->setSimVelocity(nextAngles, desiredAngularVelocities);
+
+        currentAngles = nextAngles;
+        updatePos();
+
+        // Implement Real time delay
+        rsLoop.realTimeDelay();
+        // Integrate the Simulator server 
+        simulator->server->integrateWorldThreadSafe();
+
+        operationDuration--;
+    }
+
+    desiredAngles = currentAngles;
+}
+        // Find next aqngles
