+++ continuousControl/hexapod.h
@@ -3,7 +3,7 @@
 
 #include "../modules/custom/arduinoConnection/arduinoController.h"
 #include "../modules/custom/rsTimedLoop/rsTimedLoop.h"
-#include "../modules/custom/raisimSimulatorLeg/raisimSimulator.h"
+#include "../modules/custom/raisimSimulatorFull/raisimSimulator.h"
 #include <vector>
 #include <eigen3/Eigen/Core>
 #include <eigen3/Eigen/Dense>
@@ -20,36 +20,52 @@ public:
     Eigen::MatrixXd getJacobian(int legNum) const;
 
     void updatePos();
+    Eigen::VectorXd doFK(Eigen::VectorXd angs);
     void printPos() const;
 
+    bool active;
+    int operationDuration = 0; 
+
+    // Instant set position
     void setAngs(const Eigen::VectorXd& angs);
 
+    // Velocity set position
     void moveLegToPos(const Eigen::Vector3d& desiredPos, const int legNum);
     void moveLegsToPos(const Eigen::VectorXd& desiredPos);
 
 
+    // Instant positions
     void moveToZero();
     void moveToBasic();
     void moveToOff();
     void moveToCurled();
 
+    // Velocity positions
     void stand();
+    void off();
+
+    // Velocity Path
+    void jacobianTest(const int &style);
 
     int id;
     Eigen::VectorXd pos;
     Eigen::VectorXd currentAngles;
-    Eigen::Vector3d currentAngularVelocities;
+    Eigen::VectorXd currentAngularVelocities;
+    Eigen::VectorXd desiredAngles;
+
+    RSTimedLoop& rsLoop;
+    unique_ptr<RaisimSimulator> simulator;
 
 private:
     float rsStep;
     void sendAngs();
     
     unique_ptr<ArduinoController> arduino;
-    RSTimedLoop& rsLoop;
+    
 
     bool arduinoConnected;
 
-    unique_ptr<RaisimSimulator> simulator;
+    
 
     constexpr static float coxaX = 0.044925;
     constexpr static float coxaZ = 0.01065;
